Player can no longer store more than one radio at a time. Radios should only be assigned to a player when they are the active radio. VoiceSource and all of its children have a const value stored in their .h files that holds the number of doubles required to initialize their class. added first itteration of pipe input format.

+++ Code/src/GameDataReader.cpp
@@ -81,7 +81,7 @@ Maintains game pipe connection.*/
 void GameDataReader::readFromPipe() {
 	//initialize pipe variables
 	//TODO(Ryan Lavin): make buffer of correct type - 5/28/2016
-	DWORD buffer[READER_BUFFER_SIZE];
+	double buffer[READER_BUFFER_SIZE];
 	DWORD bytesRead = 0;
 	while (true) {
 		//Bring data in from pipe
